CAOP.RAA.2024: Official Administrative Map of the Azores (CAOP 2024)

Provides the official administrative boundaries of the Azores (Região Autónoma dos Açores (RAA)) as defined in the 2024 edition of the Carta Administrativa Oficial de Portugal (CAOP), published by the Direção-Geral do Território (DGT). The package includes convenience functions to import these boundaries as 'sf' objects for spatial analysis in R. Source: https://geo2.dgterritorio.gov.pt/caop/CAOP_RAA_2024-gpkg.zip.

See also

Author

Maintainer: Ramiro Magno rmagno@pattern.institute (ORCID)

Other contributors:

Pattern Institute [copyright holder, funder]