Azorean boundary segments

boundary_segments() returns the boundary segments that separate administrative entities or borders with the Atlantic Ocean in the Azores.

boundary_segments_25N() returns the boundary segments in the meter-based projection PTRA08 / UTM zone 25N, i.e. those boundary segments in the western-most part of the archipelago: Ilha das Flores and Ilha do Corvo.

boundary_segments_26N() returns the boundary segments in the meter-based projection PTRA08 / UTM zone 26N, i.e. those boundary segments in the Central and Eastern groups of the archipelago.

Usage

boundary_segments(crs = laea_azores_proj())

boundary_segments_25N()

boundary_segments_26N()

Arguments

crs

Coordinate reference system (CRS) passed on to st_transform(). Defaults to a custom CRS centered on the Azores, see laea_azores_proj() for more details. Other possible options are "EPSG: 3035" for ETRS89-extended / LAEA Europe or "EPSG:4326" for WGS 84.

Value

A simple features (sf) object with seven fields:

id

An unique UUID identifier for the segment.

id_at_right

Identifier for the entity at the right side of the boundary.

id_at_left

Identifier for the entity at the left side of the boundary.

is_confirmed

Boundary segment whose acceptance by the parties has not yet been officially communicated.

boundary_type

Boundary type: land or coast.

boundary_level

Boundary level: administrative level.

length

Boundary segment length in kilometers.

Examples

boundary_segments()
#> Simple feature collection with 435 features and 7 fields
#> Geometry type: LINESTRING
#> Dimension:     XY
#> Bounding box:  xmin: -281288.2 ymin: -170294.3 xmax: 266022.8 ymax: 140634.8
#> Projected CRS: +proj=laea +lat_0=38.5 +lon_0=-28 +datum=WGS84 +units=m +no_defs
#> First 10 features:
#>                                      id id_at_right id_at_left is_confirmed
#> 1  79a78e76-b51a-11ee-a363-e7b43cc53be6      480203     480204        FALSE
#> 2  79a78e6e-b51a-11ee-a363-2391ce4ab53a      480102         99        FALSE
#> 3  79a78e77-b51a-11ee-a363-cba47f6db28f      480101         99        FALSE
#> 4  79a78e60-b51a-11ee-a363-5f2ec40aacd2      480105     480104        FALSE
#> 5  79a78e66-b51a-11ee-a363-df6d1e97b6f5      480106     480103        FALSE
#> 6  79a78e63-b51a-11ee-a363-976ce071c264      480105     480107        FALSE
#> 7  79a78e6a-b51a-11ee-a363-b3e954880b78      480105     480102        FALSE
#> 8  79a78e6c-b51a-11ee-a363-e71e7ba0ed77      480106     480105        FALSE
#> 9  79a78e78-b51a-11ee-a363-efd6747583e6      480203     480101        FALSE
#> 10 79a78e61-b51a-11ee-a363-2fb0cd72c871      480104         99        FALSE
#>    boundary_type boundary_level   length                       geometry
#> 1           land              5 3.933896 LINESTRING (-275904.3 11412...
#> 2          coast              3 1.358153 LINESTRING (-281037.5 10791...
#> 3          coast              3 8.515972 LINESTRING (-280613.5 10913...
#> 4           land              5 3.342975 LINESTRING (-278341.5 10159...
#> 5           land              5 5.710350 LINESTRING (-271975.5 10410...
#> 6           land              5 1.612186 LINESTRING (-277962.5 10487...
#> 7           land              5 2.752975 LINESTRING (-277707.5 10633...
#> 8           land              5 2.029468 LINESTRING (-275382.9 10724...
#> 9           land              4 1.194702 LINESTRING (-279720.5 11380...
#> 10         coast              3 8.311823 LINESTRING (-278341.5 10159...
boundary_segments_25N()
#> Simple feature collection with 31 features and 7 fields
#> Geometry type: LINESTRING
#> Dimension:     XY
#> Bounding box:  xmin: 648953 ymin: 4359450 xmax: 664413.8 ymax: 4399078
#> Projected CRS: PTRA08 / UTM zone 25N
#> First 10 features:
#>                                      id id_at_right id_at_left is_confirmed
#> 1  79a78e76-b51a-11ee-a363-e7b43cc53be6      480203     480204        FALSE
#> 2  79a78e6e-b51a-11ee-a363-2391ce4ab53a      480102         99        FALSE
#> 3  79a78e77-b51a-11ee-a363-cba47f6db28f      480101         99        FALSE
#> 4  79a78e60-b51a-11ee-a363-5f2ec40aacd2      480105     480104        FALSE
#> 5  79a78e66-b51a-11ee-a363-df6d1e97b6f5      480106     480103        FALSE
#> 6  79a78e63-b51a-11ee-a363-976ce071c264      480105     480107        FALSE
#> 7  79a78e6a-b51a-11ee-a363-b3e954880b78      480105     480102        FALSE
#> 8  79a78e6c-b51a-11ee-a363-e71e7ba0ed77      480106     480105        FALSE
#> 9  79a78e78-b51a-11ee-a363-efd6747583e6      480203     480101        FALSE
#> 10 79a78e61-b51a-11ee-a363-2fb0cd72c871      480104         99        FALSE
#>    boundary_type boundary_level   length                       geometry
#> 1           land              5 3.933896 LINESTRING (654137.1 437209...
#> 2          coast              3 1.358153 LINESTRING (649354 4365619,...
#> 3          coast              3 8.515972 LINESTRING (649710.2 436685...
#> 4           land              5 3.342975 LINESTRING (652394.7 435945...
#> 5           land              5 5.710350 LINESTRING (658613.6 436230...
#> 6           land              5 1.612186 LINESTRING (652592.2 436275...
#> 7           land              5 2.752975 LINESTRING (652766.4 436422...
#> 8           land              5 2.029468 LINESTRING (655037.9 436525...
#> 9           land              4 1.194702 LINESTRING (650344.1 437156...
#> 10         coast              3 8.311823 LINESTRING (652394.7 435945...
boundary_segments_26N()
#> Simple feature collection with 404 features and 7 fields
#> Geometry type: LINESTRING
#> Dimension:     XY
#> Bounding box:  xmin: 340225.2 ymin: 4088680 xmax: 676927.7 ymax: 4328069
#> Projected CRS: PTRA08 / UTM zone 26N
#> First 10 features:
#>                                      id id_at_right id_at_left is_confirmed
#> 1  e748fa72-b51c-11ee-a363-fb2e161626db      420204     420404         TRUE
#> 2  f108a5f0-b51b-11ee-a363-0b4741708d0a      430207     430204        FALSE
#> 3  f108a675-b51b-11ee-a363-1f0934833848      470108         99        FALSE
#> 4  f108a5f1-b51b-11ee-a363-6316321cdd58      430114     430118        FALSE
#> 5  e748fa4a-b51c-11ee-a363-d3eb99b0b6b9      420307     420312         TRUE
#> 6  e748fa9b-b51c-11ee-a363-1f313c6d8468      420308         99         TRUE
#> 7  f108a5fc-b51b-11ee-a363-ef6ebf00de6e      430110     430114        FALSE
#> 8  e748fa4d-b51c-11ee-a363-133f6865a6e9      420301     420312         TRUE
#> 9  e748fa5c-b51c-11ee-a363-cb752a0eb160      420507     420510         TRUE
#> 10 f108a679-b51b-11ee-a363-cba67cac0def      470107     470108        FALSE
#>    boundary_type boundary_level    length                       geometry
#> 1           land              4 1.0521307 LINESTRING (658386.7 418386...
#> 2           land              5 5.6535874 LINESTRING (494322.6 428420...
#> 3          coast              3 1.7670014 LINESTRING (358612.2 426702...
#> 4           land              5 7.2484001 LINESTRING (478486.8 428490...
#> 5           land              5 2.3889660 LINESTRING (617598.8 417968...
#> 6          coast              3 5.5915585 LINESTRING (619063.3 418784...
#> 7           land              5 3.7559291 LINESTRING (478486.8 428490...
#> 8           land              5 2.5385317 LINESTRING (616433.9 418126...
#> 9           land              5 0.7240755 LINESTRING (636895.5 418204...
#> 10          land              5 2.6172060 LINESTRING (358612.2 426702...