Azorean districts

districts() returns the boundaries of districts (islands) in the Azores.

districts_25N() returns the districts in the meter-based projection PTRA08 / UTM zone 25N, i.e. those districts in the western-most part of the archipelago: Ilha das Flores and Ilha do Corvo.

districts_26N() returns the districts in the meter-based projection PTRA08 / UTM zone 26N, i.e. those districts in the Central and Eastern groups of the archipelago.

Usage

districts(crs = laea_azores_proj())

districts_25N()

districts_26N()

Arguments

crs

Coordinate reference system (CRS) passed on to st_transform(). Defaults to a custom CRS centered on the Azores, see laea_azores_proj() for more details. Other possible options are "EPSG: 3035" for ETRS89-extended / LAEA Europe or "EPSG:4326" for WGS 84.

Value

A simple features (sf) object with six fields:

id

An unique identifier for the district. Follows the format [DT][MN][FR]: DT is a two-digit id for the district, MN for the municipality and FR for the district (freguesia in Portuguese).

district

Name of the civil district.

municipality

Name of municipality.

district

Name of district, coincides with the name of the island.

area

District area in hectares (ha).

perimeter

District perimeter in kilometers.

n_municipalities

Number of municipalities in the district.

n_parishes

Number of parishes in the district.

Examples

districts()
#> Simple feature collection with 9 features and 6 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -281288.2 ymin: -170294.3 xmax: 266022.8 ymax: 140634.8
#> Projected CRS: +proj=laea +lat_0=38.5 +lon_0=-28 +datum=WGS84 +units=m +no_defs
#>   id            district     area perimeter n_municipalities n_parishes
#> 1 48     Ilha das Flores 14096.22        72                2         11
#> 2 49       Ilha do Corvo  1711.45        21                1          1
#> 3 41 Ilha de Santa Maria  9688.74        78                1          5
#> 4 42  Ilha de São Miguel 74457.50       230                6         64
#> 5 43       Ilha Terceira 40026.72       126                2         30
#> 6 44    Ilha da Graciosa  6065.78        44                1          4
#> 7 45   Ilha de São Jorge 24364.78       139                2         11
#> 8 46        Ilha do Pico 44479.53       153                3         17
#> 9 47       Ilha do Faial 17305.53        80                1         13
#>                         geometry
#> 1 MULTIPOLYGON (((-278363.3 1...
#> 2 MULTIPOLYGON (((-267155.3 1...
#> 3 MULTIPOLYGON (((262934.5 -1...
#> 4 MULTIPOLYGON (((225499.7 -8...
#> 5 MULTIPOLYGON (((67086.26 17...
#> 6 MULTIPOLYGON (((4441.145 57...
#> 7 MULTIPOLYGON (((16402.23 86...
#> 8 MULTIPOLYGON (((-10785.79 -...
#> 9 MULTIPOLYGON (((-57167.91 2...
districts_25N()
#> Simple feature collection with 2 features and 6 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 648953 ymin: 4359450 xmax: 664413.8 ymax: 4399078
#> Projected CRS: PTRA08 / UTM zone 25N
#>   id        district     area perimeter n_municipalities n_parishes
#> 1 48 Ilha das Flores 14096.22        72                2         11
#> 2 49   Ilha do Corvo  1711.45        21                1          1
#>                         geometry
#> 1 MULTIPOLYGON (((652373.2 43...
#> 2 MULTIPOLYGON (((661755.7 43...
districts_26N()
#> Simple feature collection with 7 features and 6 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 340225.2 ymin: 4088680 xmax: 676927.7 ymax: 4328069
#> Projected CRS: PTRA08 / UTM zone 26N
#>   id            district     area perimeter n_municipalities n_parishes
#> 1 41 Ilha de Santa Maria  9688.74        78                1          5
#> 2 42  Ilha de São Miguel 74457.50       230                6         64
#> 3 43       Ilha Terceira 40026.72       126                2         30
#> 4 44    Ilha da Graciosa  6065.78        44                1          4
#> 5 45   Ilha de São Jorge 24364.78       139                2         11
#> 6 46        Ilha do Pico 44479.53       153                3         17
#> 7 47       Ilha do Faial 17305.53        80                1         13
#>                         geometry
#> 1 MULTIPOLYGON (((673890.3 40...
#> 2 MULTIPOLYGON (((637304.4 41...
#> 3 MULTIPOLYGON (((480046.7 42...
#> 4 MULTIPOLYGON (((417872.7 43...
#> 5 MULTIPOLYGON (((429290.7 42...
#> 6 MULTIPOLYGON (((401906.2 42...
#> 7 MULTIPOLYGON (((355680.8 42...