Azorean municipalities

municipalities() returns the boundaries of municipalities in the Azores.

municipalities_25N() returns the municipalities in the meter-based projection PTRA08 / UTM zone 25N, i.e. those municipalities in the western-most part of the archipelago: Ilha das Flores and Ilha do Corvo.

municipalities_26N() returns the municipalities in the meter-based projection PTRA08 / UTM zone 26N, i.e. those municipalities in the Central and Eastern groups of the archipelago.

Usage

municipalities(crs = laea_azores_proj())

municipalities_25N()

municipalities_26N()

Arguments

crs

Coordinate reference system (CRS) passed on to st_transform(). Defaults to a custom CRS centered on the Azores, see laea_azores_proj() for more details. Other possible options are "EPSG: 3035" for ETRS89-extended / LAEA Europe or "EPSG:4326" for WGS 84.

Value

A simple features (sf) object with six fields:

id

An unique identifier for the municipality. Follows the format [DT][MN]: DT is a two-digit id for the district and MN for the municipality.

municipality

Name of municipality.

district

Name of district, coincides with the name of the island.

area

Municipality area in hectares (ha).

perimeter

Municipality perimeter in kilometers.

n_parishes

Number of parishes in municipality.

Examples

municipalities()
#> Simple feature collection with 19 features and 6 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -281288.2 ymin: -170294.3 xmax: 266022.8 ymax: 140634.8
#> Projected CRS: +proj=laea +lat_0=38.5 +lon_0=-28 +datum=WGS84 +units=m +no_defs
#> First 10 features:
#>      id          municipality            district     area perimeter n_parishes
#> 1  4801      Lajes das Flores     Ilha das Flores  7004.79        51          7
#> 2  4802 Santa Cruz das Flores     Ilha das Flores  7091.44        52          4
#> 3  4901                 Corvo       Ilha do Corvo  1711.45        21          1
#> 4  4101         Vila do Porto Ilha de Santa Maria  9688.74        78          5
#> 5  4201                 Lagoa  Ilha de São Miguel  4559.32        45          5
#> 6  4202              Nordeste  Ilha de São Miguel 10146.80        53          9
#> 7  4203         Ponta Delgada  Ilha de São Miguel 23298.65       102         24
#> 8  4204              Povoação  Ilha de São Miguel 10640.77        64          6
#> 9  4205        Ribeira Grande  Ilha de São Miguel 18015.13       120         14
#> 10 4206  Vila Franca do Campo  Ilha de São Miguel  7796.82        58          6
#>                          geometry
#> 1  MULTIPOLYGON (((-278354.8 1...
#> 2  MULTIPOLYGON (((-270941.6 1...
#> 3  MULTIPOLYGON (((-267155.3 1...
#> 4  MULTIPOLYGON (((262934.5 -1...
#> 5  MULTIPOLYGON (((221802.1 -8...
#> 6  MULTIPOLYGON (((251756.7 -7...
#> 7  MULTIPOLYGON (((202982.7 -8...
#> 8  MULTIPOLYGON (((234835.9 -8...
#> 9  MULTIPOLYGON (((219803.1 -7...
#> 10 MULTIPOLYGON (((225482.7 -8...
municipalities_25N()
#> Simple feature collection with 3 features and 6 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 648953 ymin: 4359450 xmax: 664413.8 ymax: 4399078
#> Projected CRS: PTRA08 / UTM zone 25N
#>     id          municipality        district    area perimeter n_parishes
#> 1 4801      Lajes das Flores Ilha das Flores 7004.79        51          7
#> 2 4802 Santa Cruz das Flores Ilha das Flores 7091.44        52          4
#> 3 4901                 Corvo   Ilha do Corvo 1711.45        21          1
#>                         geometry
#> 1 MULTIPOLYGON (((652381.8 43...
#> 2 MULTIPOLYGON (((659464.4 43...
#> 3 MULTIPOLYGON (((661755.7 43...
municipalities_26N()
#> Simple feature collection with 16 features and 6 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 340225.2 ymin: 4088680 xmax: 676927.7 ymax: 4328069
#> Projected CRS: PTRA08 / UTM zone 26N
#> First 10 features:
#>      id           municipality            district     area perimeter
#> 1  4101          Vila do Porto Ilha de Santa Maria  9688.74        78
#> 2  4201                  Lagoa  Ilha de São Miguel  4559.32        45
#> 3  4202               Nordeste  Ilha de São Miguel 10146.80        53
#> 4  4203          Ponta Delgada  Ilha de São Miguel 23298.65       102
#> 5  4204               Povoação  Ilha de São Miguel 10640.77        64
#> 6  4205         Ribeira Grande  Ilha de São Miguel 18015.13       120
#> 7  4206   Vila Franca do Campo  Ilha de São Miguel  7796.82        58
#> 8  4301      Angra do Heroísmo       Ilha Terceira 23899.66       105
#> 9  4302       Praia da Vitória       Ilha Terceira 16127.06        90
#> 10 4401 Santa Cruz da Graciosa    Ilha da Graciosa  6065.78        44
#>    n_parishes                       geometry
#> 1           5 MULTIPOLYGON (((673890.3 40...
#> 2           5 MULTIPOLYGON (((633620.8 41...
#> 3           9 MULTIPOLYGON (((663651.4 41...
#> 4          24 MULTIPOLYGON (((614820.8 41...
#> 5           6 MULTIPOLYGON (((646667.4 41...
#> 6          14 MULTIPOLYGON (((631679 4180...
#> 7           6 MULTIPOLYGON (((637287.1 41...
#> 8          19 MULTIPOLYGON (((480047.1 42...
#> 9          11 MULTIPOLYGON (((486904.1 42...
#> 10          4 MULTIPOLYGON (((417872.7 43...