Azorean NUTS

nuts() returns the boundaries of the Nomenclature of Territorial Units for Statistics (NUTS).

nuts_25N() returns the NUTS boundaries in the meter-based projection PTRA08 / UTM zone 25N, i.e. the NUTS area in the western-most part of the archipelago: Ilha das Flores and Ilha do Corvo.

nuts_26N() returns the NUTS boundaries in the meter-based projection PTRA08 / UTM zone 26N, i.e. the NUTS area in the Central and Eastern groups of the archipelago.

Usage

nuts(crs = laea_azores_proj())

nuts_25N()

nuts_26N()

Arguments

crs

Coordinate reference system (CRS) passed on to st_transform(). Defaults to a custom CRS centered on the Azores, see laea_azores_proj() for more details. Other possible options are "EPSG: 3035" for ETRS89-extended / LAEA Europe or "EPSG:4326" for WGS 84.

Value

A simple features (sf) object with seven fields:

nuts_code

NUTS code.

nuts_level

NUTS level.

nuts_name

NUTS name.

area

NUTS area in hectares (ha).

perimeter

NUTS perimeter in kilometers.

n_municipalities

Number of municipalities in the NUTS.

n_parishes

Number of parishes in the NUTS.

Examples

nuts()
#> Simple feature collection with 6 features and 7 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -281288.2 ymin: -170294.3 xmax: 266022.8 ymax: 140634.8
#> Projected CRS: +proj=laea +lat_0=38.5 +lon_0=-28 +datum=WGS84 +units=m +no_defs
#>   nuts_code nuts_level                  nuts_name      area perimeter
#> 1         2          I Região Autónoma dos Açores  15807.67        93
#> 2        20         II Região Autónoma dos Açores  15807.67        93
#> 3       200        III Região Autónoma dos Açores  15807.67        93
#> 4         2          I Região Autónoma dos Açores 216388.57       849
#> 5        20         II Região Autónoma dos Açores 216388.57       849
#> 6       200        III Região Autónoma dos Açores 216388.57       849
#>   n_municipalities n_parishes                       geometry
#> 1                3         12 MULTIPOLYGON (((-278400.9 1...
#> 2                3         12 MULTIPOLYGON (((-278383.1 1...
#> 3                3         12 MULTIPOLYGON (((-278363.3 1...
#> 4               16        144 MULTIPOLYGON (((262993.7 -1...
#> 5               16        144 MULTIPOLYGON (((262971.9 -1...
#> 6               16        144 MULTIPOLYGON (((-10868.26 -...
nuts_25N()
#> Simple feature collection with 3 features and 7 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 648953 ymin: 4359450 xmax: 664413.8 ymax: 4399078
#> Projected CRS: PTRA08 / UTM zone 25N
#>   nuts_code nuts_level                  nuts_name     area perimeter
#> 1         2          I Região Autónoma dos Açores 15807.67        93
#> 2        20         II Região Autónoma dos Açores 15807.67        93
#> 3       200        III Região Autónoma dos Açores 15807.67        93
#>   n_municipalities n_parishes                       geometry
#> 1                3         12 MULTIPOLYGON (((652335 4359...
#> 2                3         12 MULTIPOLYGON (((652353 4359...
#> 3                3         12 MULTIPOLYGON (((652373.2 43...
nuts_26N()
#> Simple feature collection with 3 features and 7 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 340225.2 ymin: 4088680 xmax: 676927.7 ymax: 4328069
#> Projected CRS: PTRA08 / UTM zone 26N
#>   nuts_code nuts_level                  nuts_name     area perimeter
#> 1         2          I Região Autónoma dos Açores 216388.6       849
#> 2        20         II Região Autónoma dos Açores 216388.6       849
#> 3       200        III Região Autónoma dos Açores 216388.6       849
#>   n_municipalities n_parishes                       geometry
#> 1               16        144 MULTIPOLYGON (((673949.3 40...
#> 2               16        144 MULTIPOLYGON (((673927.6 40...
#> 3               16        144 MULTIPOLYGON (((401823.7 42...