Azorean civil administrative parishes

parishes() returns the boundaries of civil parishes in the Azores.

parishes_25N() returns the parishes in the meter-based projection PTRA08 / UTM zone 25N, i.e. those parishes in the western-most part of the archipelago: Ilha das Flores and Ilha do Corvo.

parishes_26N() returns the parishes in the meter-based projection PTRA08 / UTM zone 26N, i.e. those parishes in the Central and Eastern groups of the archipelago.

Usage

parishes(crs = laea_azores_proj())

parishes_25N()

parishes_26N()

Arguments

crs

Coordinate reference system (CRS) passed on to st_transform(). Defaults to a custom CRS centered on the Azores, see laea_azores_proj() for more details. Other possible options are "EPSG: 3035" for ETRS89-extended / LAEA Europe or "EPSG:4326" for WGS 84.

Value

A simple features (sf) object with six fields:

id

An unique identifier for the parish. Follows the format [DT][MN][FR]: DT is a two-digit id for the district, MN for the municipality and FR for the parish (freguesia in Portuguese).

parish

Name of the civil parish.

municipality

Name of municipality.

district

Name of district, coincides with the name of the island.

area

Parish area in hectares (ha).

perimeter

Parish perimeter in kilometers.

Examples

parishes()
#> Simple feature collection with 156 features and 6 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -281288.2 ymin: -170294.3 xmax: 266022.8 ymax: 140634.8
#> Projected CRS: +proj=laea +lat_0=38.5 +lon_0=-28 +datum=WGS84 +units=m +no_defs
#> First 10 features:
#>        id                parish          municipality        district    area
#> 1  480107              Mosteiro      Lajes das Flores Ilha das Flores  599.39
#> 2  480102             Fajãzinha      Lajes das Flores Ilha das Flores  615.56
#> 3  490101                 Corvo                 Corvo   Ilha do Corvo 1711.45
#> 4  480204 Santa Cruz das Flores Santa Cruz das Flores Ilha das Flores 3967.96
#> 5  480201               Caveira Santa Cruz das Flores Ilha das Flores  328.48
#> 6  480203         Ponta Delgada Santa Cruz das Flores Ilha das Flores 1765.18
#> 7  480202                Cedros Santa Cruz das Flores Ilha das Flores 1029.82
#> 8  480106                 Lomba      Lajes das Flores Ilha das Flores  987.43
#> 9  480105      Lajes das Flores      Lajes das Flores Ilha das Flores 1879.13
#> 10 480103               Fazenda      Lajes das Flores Ilha das Flores  947.36
#>    perimeter                       geometry
#> 1         12 MULTIPOLYGON (((-277962.5 1...
#> 2         13 MULTIPOLYGON (((-277707.5 1...
#> 3         21 MULTIPOLYGON (((-267155.3 1...
#> 4         37 MULTIPOLYGON (((-273740.3 1...
#> 5         10 MULTIPOLYGON (((-270883.3 1...
#> 6         25 MULTIPOLYGON (((-275904.3 1...
#> 7         18 MULTIPOLYGON (((-270573.9 1...
#> 8         18 MULTIPOLYGON (((-270883.3 1...
#> 9         24 MULTIPOLYGON (((-273020 102...
#> 10        15 MULTIPOLYGON (((-271975.5 1...
parishes_25N()
#> Simple feature collection with 12 features and 6 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 648953 ymin: 4359450 xmax: 664413.8 ymax: 4399078
#> Projected CRS: PTRA08 / UTM zone 25N
#> First 10 features:
#>        id                parish          municipality        district    area
#> 1  480107              Mosteiro      Lajes das Flores Ilha das Flores  599.39
#> 2  480102             Fajãzinha      Lajes das Flores Ilha das Flores  615.56
#> 3  490101                 Corvo                 Corvo   Ilha do Corvo 1711.45
#> 4  480204 Santa Cruz das Flores Santa Cruz das Flores Ilha das Flores 3967.96
#> 5  480201               Caveira Santa Cruz das Flores Ilha das Flores  328.48
#> 6  480203         Ponta Delgada Santa Cruz das Flores Ilha das Flores 1765.18
#> 7  480202                Cedros Santa Cruz das Flores Ilha das Flores 1029.82
#> 8  480106                 Lomba      Lajes das Flores Ilha das Flores  987.43
#> 9  480105      Lajes das Flores      Lajes das Flores Ilha das Flores 1879.13
#> 10 480103               Fazenda      Lajes das Flores Ilha das Flores  947.36
#>    perimeter                       geometry
#> 1         12 MULTIPOLYGON (((652592.2 43...
#> 2         13 MULTIPOLYGON (((652766.4 43...
#> 3         21 MULTIPOLYGON (((661755.7 43...
#> 4         37 MULTIPOLYGON (((656644.8 43...
#> 5         10 MULTIPOLYGON (((659528.2 43...
#> 6         25 MULTIPOLYGON (((654137.1 43...
#> 7         18 MULTIPOLYGON (((659542.8 43...
#> 8         18 MULTIPOLYGON (((659528.2 43...
#> 9         24 MULTIPOLYGON (((657657.6 43...
#> 10        15 MULTIPOLYGON (((658613.6 43...
parishes_26N()
#> Simple feature collection with 144 features and 6 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 340225.2 ymin: 4088680 xmax: 676927.7 ymax: 4328069
#> Projected CRS: PTRA08 / UTM zone 26N
#> First 10 features:
#>        id                           parish  municipality            district
#> 1  410101                       Almagreira Vila do Porto Ilha de Santa Maria
#> 2  410102                    Santa Bárbara Vila do Porto Ilha de Santa Maria
#> 3  410103                   Santo Espírito Vila do Porto Ilha de Santa Maria
#> 4  410104                        São Pedro Vila do Porto Ilha de Santa Maria
#> 5  410105                    Vila do Porto Vila do Porto Ilha de Santa Maria
#> 6  420101                      Água de Pau         Lagoa  Ilha de São Miguel
#> 7  420102                          Cabouco         Lagoa  Ilha de São Miguel
#> 8  420103 Lagoa (Nossa Senhora do Rosário)         Lagoa  Ilha de São Miguel
#> 9  420104               Lagoa (Santa Cruz)         Lagoa  Ilha de São Miguel
#> 10 420105                      Ribeira Chã         Lagoa  Ilha de São Miguel
#>       area perimeter                       geometry
#> 1  1121.98        17 MULTIPOLYGON (((669896.3 40...
#> 2  1527.17        25 MULTIPOLYGON (((671328.7 40...
#> 3  2667.66        33 MULTIPOLYGON (((669896.3 40...
#> 4  1817.07        27 MULTIPOLYGON (((669939.4 40...
#> 5  2554.87        32 MULTIPOLYGON (((667102.5 40...
#> 6  1745.53        29 MULTIPOLYGON (((632656.9 41...
#> 7   484.68        11 MULTIPOLYGON (((627961.1 41...
#> 8   652.12        16 MULTIPOLYGON (((625795.6 41...
#> 9  1427.27        20 MULTIPOLYGON (((627656.4 41...
#> 10  249.73        10 MULTIPOLYGON (((633636 4175...